+++ crates/engine_bibtex/src/c_api/entries.rs
@@ -6,9 +6,9 @@ use std::cell::RefCell;
 pub const ENT_STR_SIZE: usize = 250;
 
 pub struct EntryData {
-    num_entry_ints: i32,
-    num_entry_strs: i32,
-    sort_key_num: i32,
+    num_entry_ints: usize,
+    num_entry_strs: usize,
+    sort_key_num: usize,
     entry_ints: Option<XBuf<i32>>,
     entry_strs: Option<XBuf<ASCIICode>>,
 }
@@ -28,19 +28,19 @@ impl EntryData {
         self.entry_strs.as_ref().unwrap()[pos]
     }
 
-    pub fn num_ent_strs(&self) -> i32 {
+    pub fn num_ent_strs(&self) -> usize {
         self.num_entry_strs
     }
 
-    pub fn set_num_ent_strs(&mut self, val: i32) {
+    pub fn set_num_ent_strs(&mut self, val: usize) {
         self.num_entry_strs = val;
     }
 
-    pub fn sort_key_num(&self) -> i32 {
+    pub fn sort_key_num(&self) -> usize {
         self.sort_key_num
     }
 
-    pub fn set_sort_key_num(&mut self, val: i32) {
+    pub fn set_sort_key_num(&mut self, val: usize) {
         self.sort_key_num = val;
     }
 }
@@ -65,9 +65,7 @@ pub fn with_entries_mut<T>(f: impl FnOnce(&mut EntryData) -> T) -> T {
 pub extern "C" fn init_entry_ints() {
     with_entries_mut(|entries| {
         let num_cites = with_cites(|cites| cites.num_cites());
-        entries.entry_ints = Some(XBuf::new(
-            ((entries.num_entry_ints + 1) * (num_cites + 1)) as usize,
-        ));
+        entries.entry_ints = Some(XBuf::new((entries.num_entry_ints + 1) * (num_cites + 1)));
     })
 }
 
@@ -75,31 +73,30 @@ pub extern "C" fn init_entry_ints() {
 pub extern "C" fn init_entry_strs() {
     with_entries_mut(|entries| {
         let num_cites = with_cites(|cites| cites.num_cites());
-        let mut new_buf = XBuf::new(
-            (entries.num_entry_strs + 1) as usize * (num_cites + 1) as usize * (ENT_STR_SIZE + 1),
-        );
+        let mut new_buf =
+            XBuf::new((entries.num_entry_strs + 1) * (num_cites + 1) * (ENT_STR_SIZE + 1));
         new_buf.fill(127);
         entries.entry_strs = Some(new_buf);
     })
 }
 
 #[no_mangle]
-pub extern "C" fn num_ent_ints() -> i32 {
+pub extern "C" fn num_ent_ints() -> usize {
     with_entries(|entries| entries.num_entry_ints)
 }
 
 #[no_mangle]
-pub extern "C" fn set_num_ent_ints(val: i32) {
+pub extern "C" fn set_num_ent_ints(val: usize) {
     with_entries_mut(|entries| entries.num_entry_ints = val)
 }
 
 #[no_mangle]
-pub extern "C" fn num_ent_strs() -> i32 {
+pub extern "C" fn num_ent_strs() -> usize {
     with_entries(|entries| entries.num_entry_strs)
 }
 
 #[no_mangle]
-pub extern "C" fn set_num_ent_strs(val: i32) {
+pub extern "C" fn set_num_ent_strs(val: usize) {
     with_entries_mut(|entries| entries.num_entry_strs = val)
 }
